make the properties in the singer into bindable. also, should change the hue from int to float.

+++ osu.Game.Rulesets.Karaoke/Beatmaps/Metadatas/Types/ISinger.cs
@@ -9,6 +9,6 @@ namespace osu.Game.Rulesets.Karaoke.Beatmaps.Metadatas.Types
     {
         int ID { get; }
 
-        int Hue { get; }
+        float Hue { get; }
     }
 }
